Add injection to twitter
