+++ package.json
@@ -40,7 +40,8 @@
 		"semantic-ui-css": "^2.4.1",
 		"semantic-ui-react": "^0.86.0",
 		"styled-components": "^4.1.3",
-		"webextension-polyfill-ts": "^0.8.9"
+		"webextension-polyfill-ts": "^0.8.9",
+		"webpack-extension-manifest-plugin": "^0.4.3"
 	},
 	"devDependencies": {
 		"@types/chrome": "^0.0.81",
