+++ src/manifest.json
@@ -3,6 +3,16 @@
     "author": "speckle",
     "version": "0.0.1",
     "description": "__MSG_appDescription__",
+    "commands": {
+      "_execute_browser_action": {
+        "suggested_key": {
+          "windows": "Alt+Shift+S",
+          "mac": "Alt+Shift+S",
+          "chromeos": "Alt+Shift+S",
+          "linux": "Alt+Shift+S"
+        }
+      }
+    },
     "default_locale": "en",
     "icons": {
         "128": "assets/icon-128.png"
@@ -15,7 +25,9 @@
         "storage",
         "webRequest",
         "webRequestBlocking",
-        "*://*/*"
+        "*://*/*",
+        "notifications",
+        "tabs"
     ],
     "manifest_version": 2,
 
@@ -39,5 +51,8 @@
             "js": ["extension.js"],
             "matches": ["*://*/*"]
         }
+    ],
+    "web_accessible_resources": [
+        "popup.html"
     ]
 }
