+++ src/ts/contentScripts/extension/containers/SpeckleApp.tsx
@@ -3,8 +3,6 @@ import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 import { IAppState } from '../../../background/store/all'
 import { ThemeTypes } from '../../../components/styles/themes'
-import Dock from 'react-dock'
-import Vote from './Vote'
 
 interface ISpeckleApp {
   theme: ThemeTypes
@@ -17,35 +15,10 @@ interface SpeckleAppState {
 
 class SpeckleApp extends React.Component<ISpeckleApp, SpeckleAppState> {
 
-  state: SpeckleAppState = {
-    isVisible: false
-  }
-
-  buttonOnClick = () => {
-    this.setState({ isVisible: !this.state.isVisible })
-  }
-  toggleVisibility = () => {
-    this.setState({ isVisible: !this.state.isVisible })
-  }
-
   render () {
 
     return (
-      <div>
-        <h1>jhjhjhljhlhjhlk</h1>
-        <button onClick={this.buttonOnClick}>
-          Open Dock
-        </button>
-        <Dock
-          position={'right'}
-          dimMode={'transparent'}
-          defaultSize={0.4}
-          isVisible={this.state.isVisible}
-          onVisibleChange={this.toggleVisibility}
-        >
-          <Vote/>
-        </Dock>
-      </div>
+      <div/>
     )
   }
 }
